Broken apart parse method and grouped methods with similar functionality into classes
