+++ moviepy/video/io/ffmpeg_infos_parsing/parsing_handlers/chapter_handler.py
@@ -1,0 +1,27 @@
+import re
+
+
+class ChapterHandler:
+    def __init__(self, parsing_state, file_metadata):
+        self.parsing_state = parsing_state
+        self.file_metadata = file_metadata
+
+    def process_chapter_line(self, line, input_chapters):
+        if self.parsing_state._current_chapter:
+            if len(input_chapters) < self.parsing_state._current_chapter["input_number"] + 1:
+                input_chapters.append([])
+            input_chapters[self.parsing_state._current_chapter["input_number"]].append(
+                self.parsing_state._current_chapter)
+
+        chapter_data_match = re.search(
+            r"^    Chapter #(\d+):(\d+): start (\d+\.?\d+?), end (\d+\.?\d+?)",
+            line,
+        )
+        input_number, chapter_number, start, end = chapter_data_match.groups()
+
+        self.parsing_state._current_chapter = {
+            "input_number": int(input_number),
+            "chapter_number": int(chapter_number),
+            "start": float(start),
+            "end": float(end),
+        }
